Improved damping slider responsiveness, motors now initialized with correct accel rate
- The damping slider did not change the motor acceleration rate enough
at the high end of the slider. This was fixed by adding a squared
relationship between the slider position and accel rate.

- Previously, if the motors' continuous acceleration rate had been
reduced via the damping slider, if Graffik were closed and re-opened,
the motors would still move at that slower rate despite the slider
showing that it was at the fastest setting.